+++ YareZo/Engine/Graphics/Vulkan/Vk_Buffer.h
@@ -13,8 +13,8 @@ namespace Yarezo::Graphics {
         ~YzVkBuffer();
 
         void init(VkBufferUsageFlags usageFlags, VkMemoryPropertyFlags propertyFlags, size_t size, const void* data);
-        void setData(size_t size, const void* data, uint32_t offset = 0);
-        void setDynamicData(size_t size, const void* data, uint32_t offset = 0);
+        void setData(size_t size, const void* data, uint64_t offset = 0);
+        void setDynamicData(size_t size, const void* data, uint64_t offset = 0);
         void bindIndex(YzVkCommandBuffer* commandBuffer, VkIndexType type);
         void bindVertex(YzVkCommandBuffer* commandBuffer, VkDeviceSize offset);
         void flush(VkDeviceSize size = VK_WHOLE_SIZE, VkDeviceSize offset = 0);
